mutacion, bloating, cruces extra

+++ G08P2/fitness/Fitness.java
@@ -4,9 +4,9 @@ import individuo.Individuo;
 
 public interface Fitness {
 	//Obtiene el fitness del individuo
-	public int fitnessWithBloating(Individuo individuo);
+	public float fitnessWithBloating(Individuo individuo);
 	
-	public int fitness(Individuo individuo);
+	public float fitness(Individuo individuo);
 	
 	public boolean hasBloating();
 }
